ai-backend/requirements.txt
# SnapGrade AI Backend Dependencies
# Core web framework
fastapi>=0.104.1
uvicorn>=0.24.0

# Computer Vision and ML
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# Data validation and processing
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP requests for testing
requests>=2.31.0
